Jackson ObjectMapper をダイレクトフィールドアクセスに設定。ユーザガイドの加筆
