+++ src/main/java/example/Application.java
@@ -1,15 +1,11 @@
 package example;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import example.infrastructure.configuration.objectmapper.CustomObjectMapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import static springfox.documentation.builders.PathSelectors.regex;
 
 @SpringBootApplication
 @EnableSwagger2
@@ -18,20 +14,9 @@ public class Application {
         SpringApplication.run(Application.class, args);
     }
 
-    @Bean
-    public Docket newsApi() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .groupName("greetings")
-                .apiInfo(apiInfo())
-                .select()
-                .paths(regex("/greeting.*"))
-                .build();
-    }
 
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder()
-                .title("API ドキュメント by Swagger2")
-                .description("Spring REST Sample with Swagger")
-                .build();
+    @Bean
+    public ObjectMapper directFieldAccess() {
+        return new CustomObjectMapper().ofDirectFieldAccess();
     }
 }
