+++ src/main/java/example/SwaggerConfig.java
@@ -1,0 +1,29 @@
+package example;
+
+import org.springframework.context.annotation.Bean;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+
+import static springfox.documentation.builders.PathSelectors.regex;
+
+public class SwaggerConfig {
+    @Bean
+    public Docket newsApi() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .groupName("greetings")
+                .apiInfo(apiInfo())
+                .select()
+                .paths(regex("/greeting.*"))
+                .build();
+    }
+
+    private ApiInfo apiInfo() {
+        return new ApiInfoBuilder()
+                .title("API ドキュメント by Swagger2")
+                .description("Spring REST Sample with Swagger")
+                .build();
+    }
+
+}
