+++ src/main/java/example/infrastructure/configuration/objectmapper/CustomObjectMapper.java
@@ -1,0 +1,15 @@
+package example.infrastructure.configuration.objectmapper;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.PropertyAccessor;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class CustomObjectMapper {
+    public ObjectMapper ofDirectFieldAccess() {
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY);
+        mapper.setVisibility(PropertyAccessor.GETTER, JsonAutoDetect.Visibility.NONE);
+        mapper.setVisibility(PropertyAccessor.SETTER, JsonAutoDetect.Visibility.NONE);
+        return mapper;
+    }
+}
